dcs-backend-master/requirements.txt
arrow==0.15.6
click==7.1.2
cmake==3.17.3
dateparser==0.7.6
dlib==19.20.0
face-recognition==1.3.0
face-recognition-models==0.3.0
joblib==0.15.1
moment==0.10
numpy==1.18.5
Pillow==7.1.2
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
regex==2020.6.8
scikit-learn==0.23.0
scipy==1.4.1
six==1.15.0
threadpoolctl==2.1.0
times==0.7
tzlocal==2.1
opencv_contrib_python==4.5.2.54
